Added picture of ER, added nugets

+++ DataLayer/Entities/Dinosaur.cs
@@ -4,10 +4,10 @@
     {
         public int DinosaurId { get; set; } //PK
         public string DinoName { get; set; }
-        public decimal DinoWeight { get; set; } //In kg
-        public decimal DinoLenght { get; set; } //In meters
-        public decimal DinoHeight { get; set; } //In meters
-        public decimal DinoPrice { get; set; }
+        public double DinoWeight { get; set; } //In kg
+        public double DinoLenght { get; set; } //In meters
+        public double DinoHeight { get; set; } //In meters
+        public double DinoPrice { get; set; }
         public int? PromotionId { get; set; }
